Refacto removing properties.rs
